requirements.txt
autogen==0.2.0
requests==2.31.0
beautifulsoup4==4.12.2
scholarly==1.7.11
python-dotenv==1.0.0
